+++ docs/Changelog.md
@@ -34,6 +34,7 @@ This page is a brief overview of each version.
 - Better logging when SSE events have errors
 - Add support for retrieving list of models from openai and ollama providers
 - Add a Connectivity Test command and page to test whether an api is working
+- Auto install the [[AI Core Library]] initially
 
 ---
 ## 0.4.1 (2024-11-15)